issue-1691: Updating webpack.config.js and removing `@angular-builders/custom-webpack` since it's introducing more issues than fixing.
Unfortunately we cannot use @angular-builders/custom-webpack the way we thought we could.
There are assets which are "emitted" after `webpack` has completed and exited which means that we cannot reliably customize the webpack plugins (i.e. index.html is copied after webpack has exited; same goes for angular differential loading which leads to half of the assets being gzipped and the other half remain unzipped because the assets are emitted after the webpack's involvement)